ENGR00225981 csi: correct the algo when panel resolution not same as CSI output
correct the algo when panel resolution not same as CSI output. The output from CSI
is VGA, but the LCD display is WVGA. We need correct the copy algorithm if we do
not perform resize.

This patch also adds a signal handler to handle "ctrl+c" interrupt, and can exit
the program elegantly.

Signed-off-by: Robby Cai <R63905@freescale.com>